+++ grass7/raster/r.massmov/filters.h
@@ -1,12 +1,12 @@
-float gradx3(float **matrix, int row, int col, float dx, int abs);
-float grady3(float **matrix, int row, int col, float dy, int abs);
-float gradx2(float **matrix, int row, int col, float dx, int abs);
-float grady2(float **matrix, int row, int col, float dy, int abs);
-float lax(float **matrix, int row, int col, float laxfactor);
-float filter_lax(float **matrix, int row, int col, float laxfactor, float **filter_matrix,float threshold,float val);
-float gradPx2(float **matrix1, float **matrix2, float **matrix3, int row,
-		int col, float dx);
-float gradPy2(float **matrix1, float **matrix2, float **matrix3, int row,
-		int col, float dy);
-float isocell(float **matrix,int row,int col,float val);
-float filter_lax_print(float **matrix, int row, int col, float laxfactor, float **filter_matrix,float threshold,float val);
+double gradx3(double **matrix, int row, int col, double dx, int abs);
+double grady3(double **matrix, int row, int col, double dy, int abs);
+double gradx2(double **matrix, int row, int col, double dx, int abs);
+double grady2(double **matrix, int row, int col, double dy, int abs);
+double lax(double **matrix, int row, int col, double laxfactor);
+double filter_lax(double **matrix, int row, int col, double laxfactor, double **filter_matrix,double threshold,double val);
+double gradPx2(double **matrix1, double **matrix2, double **matrix3, int row,
+		int col, double dx);
+double gradPy2(double **matrix1, double **matrix2, double **matrix3, int row,
+		int col, double dy);
+double isocell(double **matrix,int row,int col,double val);
+double filter_lax_print(double **matrix, int row, int col, double laxfactor, double **filter_matrix,double threshold,double val);
